+++ app/src/main/java/com/misw4203/vinilos/presentation/viewmodel/AddAlbumToBandUiState.kt
@@ -3,8 +3,8 @@ package com.misw4203.vinilos.presentation.viewmodel
 sealed interface AddAlbumToBandUiState {
     data object Loading : AddAlbumToBandUiState
     data object Ready : AddAlbumToBandUiState
-    data class Adding(val albumId: Int) : AddAlbumToBandUiState
-    data class Error(val isNetworkError: Boolean, val albumId: Int?) : AddAlbumToBandUiState
+    data class Adding(val albumId: Long) : AddAlbumToBandUiState
+    data class Error(val isNetworkError: Boolean, val albumId: Long?) : AddAlbumToBandUiState
 }
 
 sealed interface AddAlbumToBandEvent {
